feat: 커서 기반 페이징 구현 및 QueryDsl 도입

+++ src/main/java/me/splleat/messengerproject/domain/channel/ChannelUserSettingService.java
@@ -66,15 +66,20 @@ public class ChannelUserSettingService {
 
     @Transactional
     public void leaveChannels(long userId, List<Long> channelIds) {
-        channelUserSettingRepository.deleteAllByUserIdAndChannelIdIn(userId, channelIds);
+        if (!channelIds.isEmpty()) {
+            channelUserSettingRepository.deleteAllByUserIdAndChannelIdIn(userId, channelIds);
+        }
     }
 
     @Transactional
-    public void registerIfAbsent(long userId, long channelId) {
+    public ChannelUserSetting registerIfAbsent(long userId, long channelId) {
         if (!channelUserSettingRepository.existsByUserIdAndChannelId(userId, channelId)) {
             ChannelUserSetting channelUserSetting = ChannelUserSetting.create(userId, channelId);
-            channelUserSettingRepository.save(channelUserSetting);
+            return channelUserSettingRepository.save(channelUserSetting);
         }
+
+        return channelUserSettingRepository.findByUserIdAndChannelId(userId, channelId)
+                .orElseThrow(ChannelUserSettingNotFoundException::new);
     }
 }
